backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
passlib[bcrypt]
PyJWT
requests